Running benchmarks scripts
