+++ benchmarks/run-benchmark
@@ -1,0 +1,22 @@
+#!/usr/bin/env ruby
+
+def exec(cmd)
+  puts cmd
+  abort("Error encountered, aborting") unless system(cmd)
+end
+
+unless ARGV[0]
+    $stderr.puts "ERROR: Benchmark name missing!"
+    puts
+    puts "Example: project/run-benchmark CoevalFlatMapLongLoopBenchmark"
+    exit 1
+end
+
+Dir.chdir(File.absolute_path(File.join(File.dirname(__FILE__), "..")))
+
+name = ARGV[0]
+date = Time.now.strftime("%Y-%m-%d")
+
+exec("mkdir -p benchmarks/results/#{date}")
+exec("sbt benchmarks-vnext/clean 'benchmarks-vnext/jmh:run -o ../results/#{date}/#{name}-Next.txt  -i 10 -wi 10 -f 2 -t 1 -r 1 -w 1 freestyle.rpc.benchmarks.#{name}'")
+exec("sbt benchmarks-vprev/clean 'benchmarks-vprev/jmh:run -o ../results/#{date}/#{name}-Prev.txt  -i 10 -wi 10 -f 2 -t 1 -r 1 -w 1 freestyle.rpc.benchmarks.#{name}'")
