+++ benchmarks/run-benchmarks-all
@@ -1,0 +1,10 @@
+#!/usr/bin/env ruby
+
+Dir.chdir(File.absolute_path(File.join(File.dirname(__FILE__), "..")))
+
+Dir.entries("benchmarks/shared/src/main/scala").each do |f|
+  if f =~ /^([^.]+)\.scala$/
+    name = $1
+    system("./benchmarks/run-benchmark #{name}") if name !~ /package/
+  end
+end
