+++ benchmarks/shared/src/main/scala/shared/protocols/PersonServiceAvro.scala
