+++ benchmarks/shared/src/main/scala/shared/protocols/PersonServiceAvroWithSchema.scala
